Fix wrong selector on buyChapterBatchHub and redesign batch buy UI
The Hub ABI declared endCh as uint256 instead of uint8, producing a
function selector (0x7eb7e520) that doesn't exist on the deployed
contract (real selector 0x0cdd07d2) — every single and batch chapter
purchase on the hub chain reverted because of it. Verified the fix
against the compiled contract ABI (forge inspect), not just source.

Also redesigns the batch-buy panel: segmented controls instead of
selects for track/route/payment, a compact fee stat grid instead of
seven lines of text, a toggle switch for spoke sync, and a tighter
mobile-responsive layout throughout.

+++ utils/ricoMatrixAbi.ts
@@ -136,7 +136,7 @@ export const RICO_MATRIX_HUB_ABI = [
       { name: "paymentToken", type: "address" },
       { name: "track", type: "uint8" },
       { name: "startCh", type: "uint8" },
-      { name: "endCh", type: "uint256" },
+      { name: "endCh", type: "uint8" },
     ],
     outputs: [],
   },
